Adjust format strings to work on both 32-bit and 64-bit builds
Adjust format strings to not produce Clang warnings in both 32-bit and
64-bit builds

+++ core/jni/android_view_InputEventReceiver.cpp
@@ -182,7 +182,7 @@ int NativeInputEventReceiver::handleEvent(int receiveFd, int events, void* data)
 
                 if (status == WOULD_BLOCK) {
                     if (kDebugDispatchCycle) {
-                        ALOGD("channel '%s' ~ Sent %u queued finish events; %u left.",
+                        ALOGD("channel '%s' ~ Sent %zu queued finish events; %zu left.",
                                 getInputChannelName(), i, mFinishQueue.size());
                     }
                     return 1; // keep the callback, try again later
@@ -201,7 +201,7 @@ int NativeInputEventReceiver::handleEvent(int receiveFd, int events, void* data)
             }
         }
         if (kDebugDispatchCycle) {
-            ALOGD("channel '%s' ~ Sent %u queued finish events; none left.",
+            ALOGD("channel '%s' ~ Sent %zu queued finish events; none left.",
                     getInputChannelName(), mFinishQueue.size());
         }
         mFinishQueue.clear();
@@ -218,7 +218,7 @@ status_t NativeInputEventReceiver::consumeEvents(JNIEnv* env,
         bool consumeBatches, nsecs_t frameTime, bool* outConsumedBatch) {
     if (kDebugDispatchCycle) {
         ALOGD("channel '%s' ~ Consuming input events, consumeBatches=%s, frameTime=%lld.",
-                getInputChannelName(), consumeBatches ? "true" : "false", frameTime);
+                getInputChannelName(), consumeBatches ? "true" : "false", (long long)frameTime);
     }
 
     if (consumeBatches) {
